Refactor to add command parser utils file
